requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.1

aiofiles==23.2.1  # async file operations

# Testing (dev)
pytest-asyncio==0.23.6 